@@ -5,27 +5,33 @@
   "main": "index.js",
   "dependencies": {
     "dat.gui": "^0.7.2",
-    "http-server": "^0.11.1",
+    "http-server": "^0.12.3",
     "three": "^0.97.0",
     "three-orbit-controls": "^82.1.0"
   },
   "devDependencies": {
-    "webpack": "^4.17.2",
-    "webpack-cli": "^3.1.0",
-    "webpack-dev-server": "^3.1.9",
     "@babel/core": "^7.0.0",
     "@babel/preset-env": "^7.0.0",
     "babel-loader": "^8.0.2",
-    "copy-webpack-plugin": "^4.5.2",
+    "clean-webpack-plugin": "^3.0.0",
+    "copy-webpack-plugin": "^6.0.3",
+    "css-loader": "^4.0.0",
     "eslint": "^5.5.0",
     "eslint-config-google": "^0.10.0",
     "eslint-loader": "^2.1.0",
-    "html-webpack-plugin": "^3.2.0"
+    "html-loader": "^1.1.0",
+    "html-webpack-plugin": "^3.2.0",
+    "raw-loader": "^4.0.1",
+    "style-loader": "^1.2.1",
+    "webpack": "^4.44.0",
+    "webpack-cli": "^3.3.12",
+    "webpack-dev-server": "^3.11.0"
   },
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
+    "dev": "webpack-dev-server --progress --colors",
     "watch": "webpack --watch",
-    "start": "webpack-dev-server",
+    "start": "webpack-dev-server --open",
     "build": "webpack"
   },
   "repository": {
