@@ -1,14 +1,38 @@
 const path = require('path');
+const HtmlWebpackPlugin = require('html-webpack-plugin');
+const {
+  CleanWebpackPlugin
+} = require('clean-webpack-plugin');
+const CopyWebpackPlugin = require('copy-webpack-plugin');
 
 module.exports = {
   mode: 'development',
-  entry: './src/',
+  entry: './src/index.js',
+  devtool: 'inline-source-map',
+  plugins: [
+    new CleanWebpackPlugin(),
+    new HtmlWebpackPlugin({
+      template: './src/index.html',
+      title: 'Hot Module Replacement',
+    }),
+    new CopyWebpackPlugin({
+      patterns: [{
+          from: 'src/index.html',
+          to: 'dist/index.html'
+        }
+      ]
+    })
+  ],
   output: {
-    filename: 'main.js',
+    filename: '[name].bundle.js',
     path: path.resolve(__dirname, 'dist'),
   },
   module: {
     rules: [
+      {
+        test: /\.html$/i,
+        loader: 'html-loader',
+      },
       {
         test: /\.js$/,
         exclude: /(node_modules|bower_components)/,
@@ -27,10 +51,20 @@ module.exports = {
           // eslint options (if necessary)
         }
       },
+      {
+        test: /\.css$/,
+        use: [
+          'style-loader',
+          'css-loader'
+        ],
+      }
     ]
   },
   devServer: {
     contentBase: './dist/',
+    inline: true,
+    port: 8080,
     watchContentBase: true,
-  },
+    hot: true
+  }
 };
